aaaa

+++ src/main/scala/com/gridoai/adapters/embeddingApi/Mocked.scala
@@ -4,7 +4,7 @@ import com.gridoai.domain.Embedding
 import cats.effect.IO
 
 object Mocked extends EmbeddingAPI[IO]:
-  private val mockResponse = List.range(1, 768).map(_.toDouble)
+  private val mockResponse = List.range(1, 768).map(_.toFloat)
 
   def embed(
       text: String
